// sdk/batch/batch/src/api/models.ts
export interface ImageReference {
  /** For example, Canonical or MicrosoftWindowsServer. */
  publisher?: string;
  /** For example, UbuntuServer or WindowsServer. */
  offer?: string;
  /** For example, 18.04-LTS or 2019-Datacenter. */
  sku?: string;
  /**
   * A value of 'latest' can be specified to select the latest version of an Image.
   * If omitted, the default is 'latest'.
   */
  version?: string;
  /**
   * This property is mutually exclusive with other ImageReference properties. The
   * Shared Image Gallery Image must have replicas in the same region and must be in
   * the same subscription as the Azure Batch account. If the image version is not
   * specified in the imageId, the latest version will be used. For information
   * about the firewall settings for the Batch Compute Node agent to communicate
   * with the Batch service see
   * https://docs.microsoft.com/en-us/azure/batch/batch-api-basics#virtual-network-vnet-and-firewall-configuration.
   */
  virtualMachineImageId?: string;
  /**
   * The specific version of the platform image or marketplace image used to create
   * the node. This read-only field differs from 'version' only if the value
   * specified for 'version' when the pool was created was 'latest'.
   */
  readonly exactVersion?: string;
}
// ...
export interface AccountListSupportedImagesResult {
  /** The list of supported Virtual Machine Images. */
  value?: ImageInformation[];
  /** The URL to get the next set of results. */
  nextLink?: string;
}

/**
 * A reference to the Azure Virtual Machines Marketplace Image and additional
 * information about the Image.
 */
export interface ImageInformation {
  /** The ID of the Compute Node agent SKU which the Image supports. */
  readonly nodeAgentSKUId: string;
  /**
   * A reference to an Azure Virtual Machines Marketplace Image or a Shared Image
   * Gallery Image. To get the list of all Azure Marketplace Image references
   * verified by Azure Batch, see the 'List Supported Images' operation.
   */
  imageReference: ImageReference;
  /** The type of operating system (e.g. Windows or Linux) of the Image. */
  osType: OSType;
  /**
   * Not every capability of the Image is listed. Capabilities in this list are
   * considered of special interest and are generally related to integration with
   * other features in the Azure Batch service.
   */
  capabilities?: string[];
  /**
   * The time when the Azure Batch service will stop accepting create Pool requests
   * for the Image.
   */
  batchSupportEndOfLife?: Date;
  /**
   * Whether the Azure Batch service actively verifies that the Image is compatible
   * with the associated Compute Node agent SKU.
   */
  verificationType: VerificationType;
}

/** */
/** "linux", "windows" */
export type OSType = string;
/** */
/** "verified", "unverified" */
export type VerificationType = string;

/** */
export interface PoolNodeCountsListResult {
  /** A list of Compute Node counts by Pool. */
  value?: PoolNodeCounts[];
  /** The URL to get the next set of results. */
  nextLink?: string;
}

/** The number of Compute Nodes in each state for a Pool. */
export interface PoolNodeCounts {
  /** The ID of the Pool. */
  readonly poolId: string;
  /** The number of Compute Nodes in each Compute Node state. */
  dedicated?: NodeCounts;
  /** The number of Compute Nodes in each Compute Node state. */
  lowPriority?: NodeCounts;
}

/** The number of Compute Nodes in each Compute Node state. */
export interface NodeCounts {
  /** The number of Compute Nodes in the creating state. */
  creating: number;
  /** The number of Compute Nodes in the idle state. */
  idle: number;
  /** The number of Compute Nodes in the offline state. */
  offline: number;
  /** The number of Compute Nodes in the preempted state. */
  preempted: number;
  /** The count of Compute Nodes in the rebooting state. */
  rebooting: number;
  /** The number of Compute Nodes in the reimaging state. */
  reimaging: number;
  /** The number of Compute Nodes in the running state. */
  running: number;
  /** The number of Compute Nodes in the starting state. */
  starting: number;
  /** The number of Compute Nodes in the startTaskFailed state. */
  startTaskFailed: number;
  /** The number of Compute Nodes in the leavingPool state. */
  leavingPool: number;
  /** The number of Compute Nodes in the unknown state. */
  unknown: number;
  /** The number of Compute Nodes in the unusable state. */
  unusable: number;
  /** The number of Compute Nodes in the waitingForStartTask state. */
  waitingForStartTask: number;
  /** The total number of Compute Nodes. */
  total: number;
}